+++ deepmind/engine/lua_text_level_maker.cc
@@ -314,8 +314,11 @@ bool NoOp(std::size_t, std::size_t, char,
 LuaTextLevelMaker::LuaTextLevelMaker(
     const std::string& self, const std::string& output_folder,
     bool use_local_level_cache, bool use_global_level_cache,
-    DeepMindLabLevelCacheParams level_cache_params)
-    : prng_(0), rundir_(self), output_folder_(output_folder) {
+    DeepMindLabLevelCacheParams level_cache_params, std::uint32_t mixer_seed)
+    : prng_(0),
+      mixer_seed_(mixer_seed),
+      rundir_(self),
+      output_folder_(output_folder) {
   settings_.use_local_level_cache = use_local_level_cache;
   settings_.use_global_level_cache = use_global_level_cache;
   settings_.level_cache_params = level_cache_params;
@@ -404,7 +407,7 @@ lua::NResultsOr LuaTextLevelMaker::MapFromTextLevel(lua_State* L) {
 
 
 lua::NResultsOr LuaTextLevelMaker::ViewRandomness(lua_State* L) {
-  LuaRandom::CreateObject(L, &prng_);
+  LuaRandom::CreateObject(L, &prng_, mixer_seed_);
   return 1;
 }